move GetNodeFeatures to system monitor
