+++ pkg/zos_api/network.go
@@ -13,11 +13,9 @@ import (
 func (g *ZosAPI) networkListWGPortsHandler(ctx context.Context, payload []byte) (interface{}, error) {
 	return g.networkerStub.WireguardPorts(ctx)
 }
-
 func (g *ZosAPI) networkPublicConfigGetHandler(ctx context.Context, payload []byte) (interface{}, error) {
 	return g.networkerStub.GetPublicConfig(ctx)
 }
-
 func (g *ZosAPI) networkInterfacesHandler(ctx context.Context, payload []byte) (interface{}, error) {
 	results := make(map[string][]net.IP)
 	type q struct {
@@ -44,13 +42,11 @@ func (g *ZosAPI) networkInterfacesHandler(ctx context.Context, payload []byte) (
 
 	return results, nil
 }
-
 func (g *ZosAPI) networkHasIPv6Handler(ctx context.Context, payload []byte) (interface{}, error) {
 	ipData, err := g.networkerStub.GetPublicIPv6Subnet(ctx)
 	hasIP := ipData.IP != nil && err == nil
 	return hasIP, err
 }
-
 func (g *ZosAPI) networkListPublicIPsHandler(ctx context.Context, payload []byte) (interface{}, error) {
 	return g.provisionStub.ListPublicIPs(ctx)
 }
